This PR changes DateLit represantation:

- Encode year, month and day information into an u32 field instead of keeping them separate in three different fields. The change reduces the size of DateLit from 48 bits to 32 bits.
- With the internal representation change, we changed the comparison operators' implementation. They now use one unsigned integer comparison for each method.

- Performance of the following query on TPCH data scale factor 100. 
Note: Before each query, `echo 3 > /proc/sys/vm/drop_caches && free && sync` command was run.
`SELECT EXTRACT(year FROM o_orderdate) AS orderyear FROM orders;`

|  Run    | master (ms)  | date-representation (ms) |
|----------|-----------------:|--------------------------------:|
|  1        |  34389.416    |   34850.132 
|  2        |  234.757        |   182.955 
|  3        |  221.087        |   177.373 
|  4        |  231.232        |   203.132
|  5        |  221.473        |  190.977

EDIT: Performance numbers are added.