@pateljm 2016-09-02
Nice observation and enhancement @hakanmemisoglu!