@zuyu 2016-09-03
@hakanmemisoglu My 2 cents:

 1. IMHO, such `DateLit` encode is not necessary, as inside `TypedValue` it stores in a 64-byte `union` space.
 2. This PR shares many changes with #97. Thus, you may want to work w/ @hbdeshmukh to reduce the common changes.