@hbdeshmukh 2016-09-04
Hi @zuyu 

>IMHO, such DateLit encode is not necessary, as inside TypedValue it stores in a 64-byte union space.
For the base table storage, your observation is correct. However, for evaluation of the EXTRACT and comparisons on the date, the compact representation should certainly be helpful.