@zuyu 2016-09-04
@hbdeshmukh I guess the possible performance improvements come from a) using `DateLit` instead of `DatetimeLit`, b) a better implementation of `Extract` function, but not the internal representation of `DateLit`.

@hakanmemisoglu Could you please split the changes in this PR into multiple new PRs, and then compare the performance only between two different representation of `DateLit`? Thanks!