@hakanmemisoglu 2016-09-04
@zuyu  This will be merged after @hbdeshmukh's changes. Since we are use rebasing, I don't think it will be a problem. 
For the efficiency part:
1-) Using new comparison operators obviously are faster than old counterparts using if else statements. (Of course the overall effect might not be fast as we expected in the TPCH queries.)  Thank you.
2-) No matter why TypedValue is used in query processing , the cutting the size of struct representation by 50% improves the memory usage:
 a-)  A page can contain 50% more values.
 b-) Recent x86 architectures favor using the values 32 bits or 64 bits, not 48 bits.