@zuyu 2016-09-05
> For the efficiency part:
1-) Using new comparison operators obviously are faster than old counterparts using if else statements. (Of course the overall effect might not be fast as we expected in the TPCH queries.) Thank you.

I agree there would be some improvements in the theory, but do you have the number?

> 2-) No matter why TypedValue is used in query processing , the cutting the size of struct representation by 50% improves the memory usage:
a-) A page can contain 50% more values.
b-) Recent x86 architectures favor using the values 32 bits or 64 bits, not 48 bits.

I don't think so. The existing represent of one `uint32_t` and two `uint8_t`s would take 64 bits in total due to memory alignment requirements, instead of 48 bits as what you thought. Furthermore, as we use `TypeValue` instead of `DateLit` directly, I don't see much space improvements after refactoring using one single `uint32_t`.