@hakanmemisoglu 2016-09-05
@zuyu I don't have the number. I'll post it when I have.

> I don't think so. The existing represent of one uint32_t and two uint8_ts would take 64 bits in total due to memory alignment requirements, instead of 48 bits as what you thought. 

Yes, I checked it, you are right. I did not know structs as themselves have a natural alignment. It will  be padded with 2 byte because first member.

However, this does not make my argument false, even supports it. As far as I know we store columns in natural format, then when we process them, we convert them into series of TypedValues. So keeping the sizes of structs small will help when we keep DateLit in storage memory.