@zuyu 2016-09-05
@hakanmemisoglu Again, we rarely use `DateLit` directly. On the other hand, when using `TypedValue`, any changes to `DateLit` does not save memory space due to the union alignment inside `TypedValue`. In our case, it would always be 64 bits.

So I suggest we have the numbers before merging this PR.