@hakanmemisoglu 2016-09-20
I will close this PR since an issue in the compression code should be handled first.